// src/src-button/src-button.enum.ts
export enum SrcButtonAppearance {
  primary = 'primary',
  secondary = 'secondary',
  alt1 = 'alt1',
  alt2 = 'alt2',
}

export enum SrcButtonWidth {
  auto = 'auto',
  full = 'full',
}

export enum SrcButtonState {
  default = 'default',
  success = 'success',
  selected = 'selected',
  disabled = 'disabled'
}

export enum SrcIconJustification {
  left = 'left',
  right = 'right',
}